backend/requirements.txt
requests
python-dotenv
openai>=1.30.0
fastapi
uvicorn
jinja2
